+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes
 
+#### Improvements
+- feat: enable use of replica database (delegating the choice to `DATABASES_ROUTER`) ([#359](https://github.com/jazzband/django-auditlog/pull/359))
+
+#### Important notes
+- LogEntry no longer save to same database instance is using
+
 ## 1.0.0 (2022-01-24)
 
 ### Final